philosophy_chatbot_v2.0.0/requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.3  # Use faiss-gpu if you have CUDA

# Web framework
streamlit>=1.25.0

# NLP & Text processing
nltk>=3.8
PyPDF2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional - for better performance and quantization
accelerate>=0.29.0
bitsandbytes>=0.43.0  # For 4-bit quantization

# Logging and evaluation
scikit-learn>=1.3.0